+++ backend-rust/src/indexer.rs
@@ -1799,7 +1799,7 @@ enum PreparedAccountDelegationEvent {
         account_id: i64,
         target_id:  Option<i64>,
     },
-    NoOperation,
+    RemoveBaker(BakerRemoved),
 }
 
 impl PreparedAccountDelegationEvent {
@@ -1852,8 +1852,8 @@ impl PreparedAccountDelegationEvent {
                 account_id: delegator_id.id.index.try_into()?,
             },
             DelegationEvent::BakerRemoved {
-                ..
-            } => PreparedAccountDelegationEvent::NoOperation,
+                baker_id,
+            } => PreparedAccountDelegationEvent::RemoveBaker(BakerRemoved::prepare(baker_id)?),
         };
         Ok(prepared)
     }
@@ -1917,8 +1917,83 @@ impl PreparedAccountDelegationEvent {
                 .execute(tx.as_mut())
                 .await?;
             }
-            PreparedAccountDelegationEvent::NoOperation {} => (),
+            PreparedAccountDelegationEvent::RemoveBaker(baker_removed) => {
+                baker_removed.save(tx).await?;
+            }
         }
+        Ok(())
+    }
+}
+
+/// Represents the event of a baker being removed, resulting in the delegators
+/// targeting the pool are moved to the passive pool.
+struct BakerRemoved {
+    remove_baker:    RemoveBaker,
+    move_delegators: MovePoolDelegatorsToPassivePool,
+}
+impl BakerRemoved {
+    fn prepare(baker_id: &sdk_types::BakerId) -> anyhow::Result<Self> {
+        Ok(Self {
+            remove_baker:    RemoveBaker::prepare(baker_id)?,
+            move_delegators: MovePoolDelegatorsToPassivePool::prepare(baker_id)?,
+        })
+    }
+
+    async fn save(
+        &self,
+        tx: &mut sqlx::Transaction<'static, sqlx::Postgres>,
+    ) -> anyhow::Result<()> {
+        self.remove_baker.save(tx).await?;
+        self.move_delegators.save(tx).await?;
+        Ok(())
+    }
+}
+
+/// Represents the database operation of removing baker from the baker table.
+struct RemoveBaker {
+    baker_id: i64,
+}
+impl RemoveBaker {
+    fn prepare(baker_id: &sdk_types::BakerId) -> anyhow::Result<Self> {
+        Ok(Self {
+            baker_id: baker_id.id.index.try_into()?,
+        })
+    }
+
+    async fn save(
+        &self,
+        tx: &mut sqlx::Transaction<'static, sqlx::Postgres>,
+    ) -> anyhow::Result<()> {
+        sqlx::query!("DELETE FROM bakers WHERE id=$1", self.baker_id,).execute(tx.as_mut()).await?;
+        Ok(())
+    }
+}
+
+/// Represents the database operation of moving delegators for a pool to the
+/// passive pool.
+struct MovePoolDelegatorsToPassivePool {
+    /// Baker ID of the pool to move delegators from.
+    baker_id: i64,
+}
+impl MovePoolDelegatorsToPassivePool {
+    fn prepare(baker_id: &sdk_types::BakerId) -> anyhow::Result<Self> {
+        Ok(Self {
+            baker_id: baker_id.id.index.try_into()?,
+        })
+    }
+
+    async fn save(
+        &self,
+        tx: &mut sqlx::Transaction<'static, sqlx::Postgres>,
+    ) -> anyhow::Result<()> {
+        sqlx::query!(
+            "UPDATE accounts
+             SET delegated_target_baker_id = NULL
+             WHERE delegated_target_baker_id = $1",
+            self.baker_id
+        )
+        .execute(tx.as_mut())
+        .await?;
         Ok(())
     }
 }
@@ -1949,9 +2024,7 @@ enum PreparedBakerEvent {
         staked:           i64,
         restake_earnings: bool,
     },
-    Remove {
-        baker_id: i64,
-    },
+    Remove(BakerRemoved),
     StakeIncrease {
         baker_id: i64,
         staked:   i64,
@@ -1965,8 +2038,10 @@ enum PreparedBakerEvent {
         restake_earnings: bool,
     },
     SetOpenStatus {
-        baker_id:    i64,
-        open_status: BakerPoolOpenStatus,
+        baker_id:        i64,
+        open_status:     BakerPoolOpenStatus,
+        /// When set to ClosedForAll move delegators to passive pool.
+        move_delegators: Option<MovePoolDelegatorsToPassivePool>,
     },
     SetMetadataUrl {
         baker_id:     i64,
@@ -2008,9 +2083,7 @@ impl PreparedBakerEvent {
             },
             BakerEvent::BakerRemoved {
                 baker_id,
-            } => PreparedBakerEvent::Remove {
-                baker_id: baker_id.id.index.try_into()?,
-            },
+            } => PreparedBakerEvent::Remove(BakerRemoved::prepare(baker_id)?),
             BakerEvent::BakerStakeIncreased {
                 baker_id,
                 new_stake,
@@ -2038,10 +2111,19 @@ impl PreparedBakerEvent {
             BakerEvent::BakerSetOpenStatus {
                 baker_id,
                 open_status,
-            } => PreparedBakerEvent::SetOpenStatus {
-                baker_id:    baker_id.id.index.try_into()?,
-                open_status: open_status.to_owned().into(),
-            },
+            } => {
+                let open_status = open_status.to_owned().into();
+                let move_delegators = if matches!(open_status, BakerPoolOpenStatus::ClosedForAll) {
+                    Some(MovePoolDelegatorsToPassivePool::prepare(baker_id)?)
+                } else {
+                    None
+                };
+                PreparedBakerEvent::SetOpenStatus {
+                    baker_id: baker_id.id.index.try_into()?,
+                    open_status,
+                    move_delegators,
+                }
+            }
             BakerEvent::BakerSetMetadataURL {
                 baker_id,
                 metadata_url,
@@ -2114,12 +2196,8 @@ impl PreparedBakerEvent {
                 .execute(tx.as_mut())
                 .await?;
             }
-            PreparedBakerEvent::Remove {
-                baker_id,
-            } => {
-                sqlx::query!("DELETE FROM bakers WHERE id=$1", baker_id,)
-                    .execute(tx.as_mut())
-                    .await?;
+            PreparedBakerEvent::Remove(baker_removed) => {
+                baker_removed.save(tx).await?;
             }
             PreparedBakerEvent::StakeIncrease {
                 baker_id,
@@ -2152,6 +2230,7 @@ impl PreparedBakerEvent {
             PreparedBakerEvent::SetOpenStatus {
                 baker_id,
                 open_status,
+                move_delegators,
             } => {
                 sqlx::query!(
                     "UPDATE bakers SET open_status = $2 WHERE id=$1",
@@ -2160,6 +2239,9 @@ impl PreparedBakerEvent {
                 )
                 .execute(tx.as_mut())
                 .await?;
+                if let Some(move_operation) = move_delegators {
+                    move_operation.save(tx).await?;
+                }
             }
             PreparedBakerEvent::SetMetadataUrl {
                 baker_id,