+++ backend-rust/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+Database schema version: 4
+
+### Fixed
+
+- Add database migration fixing invalid delegator state, caused by validator/baker getting removed or changing status to 'ClosedForAll' without moving delegators to the passive pool.
+- Fixed indexer missing handling of moving delegators as pool got removed or closed.
+- Fixed indexer missing handling of event of baker switching directly to delegation.
+
 ## [0.1.24] - 2025-02-12
 
 ### Fixed